archivos_completos/requirements.txt
Django>=4.2.0
Pillow>=10.0.0
python-dateutil>=2.8.2
python-docx>=0.8.11
